Part 4:  Add an interface between the StudentRepository and the StudentModel and call it IStudentInterface.
Analyze the code for the Student Repository and determine which method(s) the repository uses in the Student Model and place those in the interface.
In StudentRepository, change all references to StudentModel to IStudentInterface, except for calling the Student Model constructors.
In iStudentCRUDInterface, replace all references to StudentModel to IStudentInterface
Change the StudentModel variables to IStudentInterface in StudentController.cs.
Change the StudentModel model to IStudentInterface in all the views (Create.cshtml, Delete.cshtml, Details.cshtml, Edit.cshtml, and Index.cshtml).
Note that the StudentController and the views need get/set access to the Id, Name and Credits properties through the I StudentInterface.